rename LastMessages to MessagesTable
